aspect fixes

+++ dinput8/fix_aspect.h
@@ -11,7 +11,7 @@ struct Patch;
 void RegisterPatch_Aspect();
 bool ApplyPatch_Aspect(Patch* patch);
 
-extern float fovMultiplier;
+extern double fovMultiplier;
 
 int __cdecl Hook_Atoi(const char* string);
